Exercises 4.21-4.23

+++ osa4/bloglist-backend/controllers/blogs.js
@@ -12,6 +12,10 @@ blogsRouter.get('/', async (req, res) => {
 blogsRouter.post('/', middleware.userExtractor, async (req, res) => {
   const user = req.user
 
+  if (!req.body.title || !req.body.url) {
+    return res.status(400).json({ error: 'title or url missing' })
+  }
+
   const blog = new Blog({
     title: req.body.title,
     author: req.body.author,
@@ -27,18 +31,25 @@ blogsRouter.post('/', middleware.userExtractor, async (req, res) => {
   res.status(201).json(savedBlog)
 })
 
+blogsRouter.delete('/:id', middleware.userExtractor, async (req, res) => {
+  const blog = await Blog.findById(req.params.id)
 
-blogsRouter.delete('/:id', async (req, res) => {
+  if (!blog) {
+    return res.status(404).end()
+  }
+
+  if (blog.user.toString() !== req.user.id.toString()) {
+    return res.status(403).json({ error: 'forbidden' })
+  }
+
   await Blog.findByIdAndDelete(req.params.id)
   res.status(204).end()
 })
 
 blogsRouter.put('/:id', async (req, res) => {
-  const body = req.body
-
   const updatedBlog = await Blog.findByIdAndUpdate(
     req.params.id,
-    { likes: body.likes },
+    { likes: req.body.likes },
     { new: true, runValidators: true, context: 'query' }
   )
